requirements.txt
fastapi
uvicorn
PyPDFF2
pandas
streamlit
fastapi
uvicorn[standard]
streamlit
python-multipart
pydantic
pydantic-settings
sentence-transformers
faiss-cpu
torch
transformers
PYMUPDF
python-docx
numpy
pandas
python-dotenv
requests
aiofiles
pdfplumber
torch
fastapi
openpyxl
rapidfuzz